+++ packages/models/src/config/behaviors/api.ts
@@ -5,6 +5,7 @@ import { apiServerBehaviors } from "../../server/Server/behaviors/index.js";
 import { apiCustomerBehaviors } from "../../customer/Customer/behaviors/index.js";
 import { apiIngressBehaviors } from "../../domain/Ingress/behaviors/index.js";
 import { apiAppInstallationBehaviors } from "../../app/AppInstallation/behaviors/index.js";
+import { addUrlTagToProvideReactCache } from "../../react/asyncResourceInvalidation.js";
 
 class ApiSetupState {
   private _client: MittwaldAPIV2Client | undefined;
@@ -16,6 +17,8 @@ class ApiSetupState {
       );
     }
     this._client = client;
+    this._client.defaultRequestOptions.onBeforeRequest =
+      addUrlTagToProvideReactCache;
 
     config.behaviors.project = apiProjectBehaviors(client);
     config.behaviors.server = apiServerBehaviors(client);