+++ packages/models/src/domain/Ingress/types.ts
@@ -1,8 +1,15 @@
 import { MittwaldAPIV2 } from "@mittwald/api-client";
+import { Project } from "../../project/index.js";
 
-export interface IngressListQuery {
-  projectId?: string;
-}
+export type IngressListQueryData =
+  MittwaldAPIV2.Paths.V2Ingresses.Get.Parameters.Query;
+
+export type IngressListQueryModelData = Omit<
+  IngressListQueryData,
+  "projectId"
+> & {
+  project?: Project;
+};
 
 export type IngressData =
   MittwaldAPIV2.Operations.IngressGetIngress.ResponseData;